+++ app/Http/Controllers/Workflow/WorkflowController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\Workflow\WorkflowFormRequest;
 use App\Models\ReferenceData\ReferenceData;
 use App\Models\Workflow\Workflow;
 use Exception;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class WorkflowController extends Controller
@@ -16,6 +17,7 @@ class WorkflowController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Workflow::class);
         $workflows = Workflow::with('country', 'priceplan')->get();
 
         return Inertia::render('Workflow/WorkflowIndex', [
@@ -28,6 +30,7 @@ class WorkflowController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Workflow::class);
         $status = ReferenceData::fullData()
             ->where('domain', 'Workflow')
             ->where('parameter', 'Status')
@@ -59,6 +62,7 @@ class WorkflowController extends Controller
      */
     public function show(string $id)
     {
+        Gate::authorize('view', Workflow::class);
         $workflow = Workflow::with('country', 'priceplan', 'workflowModules.workflowItems')->findOrFail($id);
 
         return Inertia::render('Workflow/WorkflowShow', [
@@ -71,6 +75,7 @@ class WorkflowController extends Controller
      */
     public function edit(string $id)
     {
+        Gate::authorize('update', Workflow::class);
         $workflow = Workflow::with('country', 'priceplan')->find($id);
         $status = ReferenceData::fullData()
             ->where('domain', 'Workflow')
@@ -104,6 +109,7 @@ class WorkflowController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize('delete', Workflow::class);
         try {
             Workflow::find($id)->delete();
         } catch (Exception $e) {
